requirements.txt
aiosqlite
loguru
pyfiglet
pystyle
PyYAML
g4f
